+++ framework/db-http/src/main/java/is/codion/framework/db/http/JsonHttpEntityConnection.java
@@ -223,7 +223,7 @@ final class JsonHttpEntityConnection extends AbstractHttpEntityConnection {
   }
 
   @Override
-  public Collection<Entity.Key> insert(Collection<? extends Entity> entities) throws DatabaseException {
+  public Collection<Entity.Key> insert(Collection<Entity> entities) throws DatabaseException {
     Objects.requireNonNull(entities);
     try {
       synchronized (this.entities) {
@@ -241,7 +241,7 @@ final class JsonHttpEntityConnection extends AbstractHttpEntityConnection {
   }
 
   @Override
-  public Collection<Entity> insertSelect(Collection<? extends Entity> entities) throws DatabaseException {
+  public Collection<Entity> insertSelect(Collection<Entity> entities) throws DatabaseException {
     Objects.requireNonNull(entities);
     try {
       synchronized (this.entities) {
@@ -269,7 +269,7 @@ final class JsonHttpEntityConnection extends AbstractHttpEntityConnection {
   }
 
   @Override
-  public void update(Collection<? extends Entity> entities) throws DatabaseException {
+  public void update(Collection<Entity> entities) throws DatabaseException {
     Objects.requireNonNull(entities);
     try {
       synchronized (this.entities) {
@@ -286,7 +286,7 @@ final class JsonHttpEntityConnection extends AbstractHttpEntityConnection {
   }
 
   @Override
-  public Collection<Entity> updateSelect(Collection<? extends Entity> entities) throws DatabaseException {
+  public Collection<Entity> updateSelect(Collection<Entity> entities) throws DatabaseException {
     Objects.requireNonNull(entities);
     try {
       synchronized (this.entities) {
@@ -462,7 +462,7 @@ final class JsonHttpEntityConnection extends AbstractHttpEntityConnection {
   }
 
   @Override
-  public Map<EntityType, Collection<Entity>> dependencies(Collection<? extends Entity> entities) throws DatabaseException {
+  public Map<EntityType, Collection<Entity>> dependencies(Collection<Entity> entities) throws DatabaseException {
     Objects.requireNonNull(entities, "entities");
     try {
       Map<EntityType, Collection<Entity>> dependencies = new HashMap<>();